Changes
for providers, fixes for missing thumbnail in identify.
