+++ RetroGamesCore.cs
@@ -1,13 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
 using MediaBrowser.Controller.Entities;
 using MediaBrowser.Controller.Library;
-using MediaBrowser.Controller.Providers;
 using MediaBrowser.Controller.Resolvers;
 using MediaBrowser.Model.Entities;
-using MediaBrowser.Model.Providers;
-using Microsoft.Extensions.Logging;
 
 namespace JellyEmu
 {
+    /// <summary>
+    /// Tells Jellyfin how to identify ROM files and folders in the library scanner.
+    /// </summary>
     public class RomResolver : IItemResolver
     {
         private readonly PlatformResolver _platformResolver;
@@ -38,6 +41,7 @@ namespace JellyEmu
                         var binPath     = CueParser.GetFirstBinPath(cueFiles[0]);
                         var consoleTag  = _platformResolver.Resolve(cuePath);
                         var regionTag   = PlatformResolver.ResolveRegion(cuePath);
+                        
                         // Use the folder name as the display name — it's usually cleaner
                         // than the cue filename, and is what the user sees in the library.
                         var displayName = PlatformResolver.CleanDisplayName(
@@ -118,99 +122,6 @@ namespace JellyEmu
             }
 
             return null;
-        }
-    }
-
-    public class RomMetadataProvider : ILocalMetadataProvider<Book>, IRemoteImageProvider
-    {
-        private readonly ILogger<RomMetadataProvider> _logger;
-        private readonly PlatformResolver _platformResolver;
-
-        public RomMetadataProvider(ILogger<RomMetadataProvider> logger, PlatformResolver platformResolver)
-        {
-            _logger = logger;
-            _platformResolver = platformResolver;
-        }
-
-        public string Name => "Retro Games Local Metadata";
-
-        public Task<MetadataResult<Book>> GetMetadata(
-            ItemInfo info,
-            IDirectoryService directoryService,
-            CancellationToken cancellationToken)
-        {
-            var result = new MetadataResult<Book>();
-
-            if (!RomExtensions.IsRomPath(info.Path))
-                return Task.FromResult(result);
-
-            var effectivePath = RomExtensions.EffectiveRomPath(info.Path);
-            var nfoPath = Path.ChangeExtension(effectivePath, ".nfo");
-
-            if (File.Exists(nfoPath))
-            {
-                var consoleTag = _platformResolver.Resolve(effectivePath);
-
-                result.HasMetadata = true;
-                result.Item = new Book
-                {
-                    Overview = "Parsed successfully from local .nfo file!",
-                    PremiereDate = new DateTime(1990, 1, 1),
-                    Tags = new[] { "Game", consoleTag }
-                };
-            }
-
-            return Task.FromResult(result);
-        }
-
-        public bool Supports(BaseItem item)
-        {
-            if (item is not Book) return false;
-            if (string.IsNullOrEmpty(item.Path)) return true;
-            return RomExtensions.IsRomPath(item.Path);
-        }
-
-        public IEnumerable<ImageType> GetSupportedImages(BaseItem item) => new[] { ImageType.Primary, ImageType.Backdrop };
-
-        public async Task<IEnumerable<RemoteImageInfo>> GetImages(BaseItem item, CancellationToken cancellationToken)
-        {
-            var list = new List<RemoteImageInfo>();
-            if (!string.IsNullOrEmpty(item.Path))
-            {
-                var effectivePath = RomExtensions.EffectiveRomPath(item.Path);
-                var dir = Path.GetDirectoryName(effectivePath);
-                var baseName = Path.GetFileNameWithoutExtension(effectivePath);
-                var possible = Path.Combine(dir ?? string.Empty, baseName + ".jpg");
-                if (File.Exists(possible))
-                {
-                    list.Add(new RemoteImageInfo { ProviderName = Name, Type = ImageType.Primary, Url = new Uri(possible).AbsoluteUri });
-                }
-                var possiblePng = Path.ChangeExtension(possible, ".png");
-                if (File.Exists(possiblePng))
-                {
-                    list.Add(new RemoteImageInfo { ProviderName = Name, Type = ImageType.Primary, Url = new Uri(possiblePng).AbsoluteUri });
-                }
-            }
-            return list;
-        }
-
-        public Task<System.Net.Http.HttpResponseMessage> GetImageResponse(string url, CancellationToken cancellationToken)
-        {
-            if (!string.IsNullOrEmpty(url) && url.StartsWith("file://", StringComparison.OrdinalIgnoreCase))
-            {
-                var localPath = new Uri(url).LocalPath;
-                if (File.Exists(localPath))
-                {
-                    var response = new System.Net.Http.HttpResponseMessage(System.Net.HttpStatusCode.OK);
-                    var stream = File.OpenRead(localPath);
-                    response.Content = new System.Net.Http.StreamContent(stream);
-                    var ext = Path.GetExtension(localPath).ToLowerInvariant();
-                    var mime = ext == ".png" ? "image/png" : "image/jpeg";
-                    response.Content.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue(mime);
-                    return Task.FromResult(response);
-                }
-            }
-            return Task.FromResult<System.Net.Http.HttpResponseMessage>(null!);
         }
     }
 }
